Move WGAddress struct to separated file
